Move cyclic actions from actions.ts
