+++ app/frontend/helpers/debounceEvent.ts
@@ -1,0 +1,10 @@
+export default (callback, time) => {
+  let interval
+  return (...args) => {
+    clearTimeout(interval)
+    interval = setTimeout(() => {
+      interval = null
+      callback(...args)
+    }, time)
+  }
+}
